Get parent directory of default powershell location
The $profile variable path leaf is the name of the file. The destination needs to be the parent directory of the file because the new profile file will be migrated to the parent directory. Since the leaf of the path is a file, a file is not a container to hold other files.